Broadcast using spanning tree
Assuming a distributed system in which nodes are arranged in a certain
topology (specified in a
configuration file). Built a spanning tree using a distributed
algorithm. Once the spanning tree construction completes, each node
should know which subset of its neighbors are also its tree neighbor.
Using the spanning tree constructed above implemented a broadcast
service that allows any node
to send a message to all nodes in the system. The broadcast service
informs the source node
of the completion of the broadcast operation. Assumption is made that
only one broadcast
operation can occur at any time.